Modifying FindOpenSSL to make direct use of set_target_properties
This is required to get this building with CMake 3.1.  CMake 3.1 changes the way that link libraries are specified; a single target can only have one link library, not a list of them, and if a list is required then synthetic targets are needed to represent them on a per-entry basis.